+++ azplugins/__init__.py
@@ -3,6 +3,7 @@
 
 # Maintainer: mphoward / All developers are free to add new modules
 
+import analyze
 import evaporate
 import pair
 import update